requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.8
SQLAlchemy==2.0.15
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
python-multipart==0.0.6
